moved lincoln_vehicle_factory to vehicle/lincoln/ folder for better code organization
