+++ modules/canbus/vehicle/vehicle_factory.cc
@@ -15,7 +15,7 @@
  *****************************************************************************/
 
 #include "modules/canbus/proto/vehicle_parameter.pb.h"
-#include "modules/canbus/vehicle/lincoln_vehicle_factory.h"
+#include "modules/canbus/vehicle/lincoln/lincoln_vehicle_factory.h"
 #include "modules/canbus/vehicle/vehicle_factory.h"
 
 namespace apollo {
